Move skiko.renderApi to SkikoProperties
